+++ sdk/resourcemanager/healthcareapis/armhealthcareapis/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
 
+## 2.1.0 (2024-01-12)
+### Features Added
+
+- New enum type `SmartDataActions` with values `SmartDataActionsRead`
+- New struct `SmartIdentityProviderApplication`
+- New struct `SmartIdentityProviderConfiguration`
+- New field `SmartIdentityProviders` in struct `FhirServiceAuthenticationConfiguration`
+
+
 ## 2.0.0 (2023-12-22)
 ### Breaking Changes